align data

+++ include/Operations/Operation_DigitalPinWrite.h
@@ -17,7 +17,7 @@ namespace OperationArchitecture
 
 		void Execute(bool x) override;
 
-		static IOperationBase *Create(const void *config, unsigned int &sizeOut, const EmbeddedIOServices::EmbeddedIOServiceCollection *embeddedIOServiceCollection);
+		static IOperationBase *Create(const void *config, size_t &sizeOut, const EmbeddedIOServices::EmbeddedIOServiceCollection *embeddedIOServiceCollection);
 	};
 }
 #endif
