+++ include/Operations/Operation_FrequencyPinRead.h
@@ -16,7 +16,7 @@ namespace OperationArchitecture
 
 		float Execute() override;
 
-		static IOperationBase *Create(const void *config, unsigned int &sizeOut, const EmbeddedIOServices::EmbeddedIOServiceCollection *embeddedIOServiceCollection);
+		static IOperationBase *Create(const void *config, size_t &sizeOut, const EmbeddedIOServices::EmbeddedIOServiceCollection *embeddedIOServiceCollection);
 	};
 }
 #endif
