+++ include/Operations/Operation_PwmPinWrite.h
@@ -16,7 +16,7 @@ namespace OperationArchitecture
 
 		void Execute(float period, float pulseWidth) override;
 
-		static IOperationBase *Create(const void *config, unsigned int &sizeOut, const EmbeddedIOServices::EmbeddedIOServiceCollection *embeddedIOServiceCollection);
+		static IOperationBase *Create(const void *config, size_t &sizeOut, const EmbeddedIOServices::EmbeddedIOServiceCollection *embeddedIOServiceCollection);
 	};
 }
 #endif
