+++ src/Operations/Operation_AnalogPinRead.cpp
@@ -19,7 +19,7 @@ namespace OperationArchitecture
 		return _analogService->ReadPin(_pin);
 	}
 
-	IOperationBase *Operation_AnalogPinRead::Create(const void *config, unsigned int &sizeOut, const EmbeddedIOServiceCollection *embeddedIOServiceCollection)
+	IOperationBase *Operation_AnalogPinRead::Create(const void *config, size_t  &sizeOut, const EmbeddedIOServiceCollection *embeddedIOServiceCollection)
 	{
 		return new Operation_AnalogPinRead(embeddedIOServiceCollection->AnalogService, Config::CastAndOffset<uint16_t>(config, sizeOut));
 	}
