+++ src/Operations/Operation_DigitalPinRead.cpp
@@ -22,7 +22,7 @@ namespace OperationArchitecture
 		return _digitalService->ReadPin(_pin);
 	}
 
-	IOperationBase *Operation_DigitalPinRead::Create(const void *config, unsigned int &sizeOut, const EmbeddedIOServiceCollection *embeddedIOServiceCollection)
+	IOperationBase *Operation_DigitalPinRead::Create(const void *config, size_t  &sizeOut, const EmbeddedIOServiceCollection *embeddedIOServiceCollection)
 	{
 		const uint16_t pin = Config::CastAndOffset<uint16_t>(config, sizeOut);
 		const bool inverted = Config::CastAndOffset<bool>(config, sizeOut);
