+++ src/Operations/Operation_DigitalPinRecord.cpp
@@ -55,7 +55,7 @@ namespace OperationArchitecture
 		_record.Last = last;
 	}
 
-	IOperationBase *Operation_DigitalPinRecord::Create(const void *config, unsigned int &sizeOut, const EmbeddedIOServiceCollection *embeddedIOServiceCollection)
+	IOperationBase *Operation_DigitalPinRecord::Create(const void *config, size_t  &sizeOut, const EmbeddedIOServiceCollection *embeddedIOServiceCollection)
 	{
 		const uint16_t pin = Config::CastAndOffset<uint16_t>(config, sizeOut);
 		const bool inverted = Config::CastAndOffset<bool>(config, sizeOut);
