+++ src/Operations/Operation_PwmPinWrite.cpp
@@ -20,7 +20,7 @@ namespace OperationArchitecture
 		_pwmService->WritePin(_pin, { period, pulseWidth });
 	}
 
-	IOperationBase *Operation_PwmPinWrite::Create(const void *config, unsigned int &sizeOut, const EmbeddedIOServiceCollection *embeddedIOServiceCollection)
+	IOperationBase *Operation_PwmPinWrite::Create(const void *config, size_t  &sizeOut, const EmbeddedIOServiceCollection *embeddedIOServiceCollection)
 	{
 		const uint16_t pin = Config::CastAndOffset<uint16_t>(config, sizeOut);
 		const uint16_t minFrequency = Config::CastAndOffset<uint16_t>(config, sizeOut);
